ci: speed up test workflow
- Drop Xdebug from the tests job (coverage: none) — it was enabled but no
  coverage was collected, taxing every test 2-5x.
- Run the Pest suite with --parallel (adds brianium/paratest to require-dev,
  already present transitively via Pest).
- Split Larastan (types:check) into its own job so it runs alongside tests
  instead of blocking them.
- Cache composer (ramsey/composer-install) and bun (~/.bun/install/cache)
  deps in both tests and lint workflows.
- Make three test helpers parallel-safe by moving them into tests/Pest.php
  (loaded by every paratest worker): issuedKey, transparentPng, publishTarget,
  bindConnector. Previously defined in one test file and used across others,
  which only works under serial execution.
