+++ composer.json
@@ -30,6 +30,7 @@
         "sentry/sentry-laravel": "^4.26"
     },
     "require-dev": {
+        "brianium/paratest": "^7.20",
         "driftingly/rector-laravel": "^2.5",
         "fakerphp/faker": "^1.24",
         "larastan/larastan": "^3.9",
