+++ tests/Feature/Api/ApiAuthTest.php
@@ -1,34 +1,7 @@
 <?php
 
 use App\Models\ApiKey;
-use App\Models\User;
-use App\Models\Workspace;
 use App\Services\Api\ApiKeyManager;
-use Illuminate\Support\Facades\Artisan;
-use Laravel\Passport\Client;
-
-/**
- * Issue a real key and return [User, Workspace, plaintextToken]. The user is a
- * member of the workspace.
- *
- * @return array{0: User, 1: Workspace, 2: string}
- */
-function issuedKey(string $scope = 'write'): array
-{
-    if (! file_exists(storage_path('oauth-private.key'))) {
-        Artisan::call('passport:keys', ['--no-interaction' => true]);
-    }
-
-    Client::factory()->asPersonalAccessTokenClient()->create(['provider' => 'users']);
-
-    $user = User::factory()->create();
-    $workspace = Workspace::factory()->create();
-    $workspace->members()->create(['user_id' => $user->id, 'role' => 'admin']);
-
-    [, $plain] = app(ApiKeyManager::class)->issue($workspace, $user, 'test', $scope, null);
-
-    return [$user, $workspace, $plain];
-}
 
 test('a request without a token is 401', function () {
     $this->getJson('/api/v1/connected-accounts')->assertUnauthorized();
